adding unit tests for question and partion function

+++ src/question.h
@@ -5,11 +5,11 @@
 class question {
     private:
         const uint32_t feature;
-        const uint32_t val;
+        const double val;
         const feature_t type;
 
     public:
-        question(uint32_t feature, uint32_t val, feature_t type) :
+        question(uint32_t feature, double val, feature_t type) :
             feature(feature), val(val), type(type) {}
 
         bool eval(const row_t& row) const;
